Delete old Tooltip component. Move V2 as the official Tooltip component
