--- src/components/Tooltip/docs/Popper.md
@@ -1,7 +1,0 @@
-# Popper
-
-A Toolbar.Popper is an internal [Tooltip](./Tooltip.md) component, that wraps [Pop.Popper](../Pop/docs/Popper).
-
-## Example
-
-This component is powered by [Pop](../Pop). After additional props, check out [Pop's documentation](../Pop).
