+++ src/components/Tooltip/index.js
@@ -1,4 +1,5 @@
 import Tooltip from './Tooltip'
+
 export { TooltipContext } from './Tooltip'
 
 export default Tooltip
