--- src/components/TooltipV2/index.js
@@ -1,5 +1,0 @@
-import Tooltip from './Tooltip'
-
-export { TooltipContext } from './Tooltip'
-
-export default Tooltip
